// server/src/types/enums.ts
export enum ERROR_MESSAGES {
  INCORRECT_EMAIL_OR_PASSWORD = "INCORRECT_EMAIL_OR_PASSWORD",
  INCORRECT_EMAIL = "INCORRECT_EMAIL",
  INCORRECT_REGISTER_DATA = "INCORRECT_REGISTER_DATA",
  DUPLICATE = "DUPLICATE",
  NOT_LOGGED_IN = "NOT_LOGGED_IN",
  SERVER_ERROR = "SERVER_ERROR",
  NO_SECRET_KEY_DEFINED = "NO_SECRET_KEY_DEFINED",
  INCORRECT_TOKEN = "INCORRECT_TOKEN",
  ERROR_ON_FINDING_POST = "ERROR_ON_FINDING_POST",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INCORRECT_POST_INPUTS = "INCORRECT_POST_INPUTS",
}